requirements.txt
# Core requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0  # Added this
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
sse-starlette>=1.8.2
loguru>=0.7.2

# Testing
pytest>=7.4.3
httpx>=0.27.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Logging
loguru>=0.7.2